hr_system/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
Js2Py==0.70
nodejs==0.1.1
npm==0.1.1
optional-django==0.1.0
pyjsparser==2.7.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
